backend/requirements.txt
PyJWT==2.7.0
pylint==2.17.4
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
reportlab==4.0.4
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
wrapt==1.15.0
asgiref==3.6.0
astroid==2.15.5
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
Django==3.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
drf-yasg==1.21.6
flake8==6.0.0
gunicorn==20.0.4
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
Pillow==9.4.0
platformdirs==3.5.3
psycopg2-binary==2.8.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
